pyproject.toml
[tool.poetry]
name = "tfmini"
version = "0.2.0"
description = "A driver for the TFmini LiDAR sold by Sparkfun"
authors = ["walchko <walchko@users.noreply.github.com>"]
readme = "readme.rst"
license = "MIT"
homepage = "https://pypi.org/project/tfmini/"
repository = 'http://github.com/MomsFriendlyRobotCompany/tfmini'
# documentation = "http://..."
keywords = ['robot', 'pi', 'serial', 'sensor', 'range', 'ranging', 'robotics', 'tf']
classifiers = [
    'Development Status :: 4 - Beta',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3.6',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Topic :: Software Development :: Libraries',
    'Topic :: Software Development :: Libraries :: Python Modules',
    'Topic :: Software Development :: Libraries :: Application Frameworks'
]

[tool.poetry.dependencies]
python = ">=3.6"
pyserial = "*"
importlib-metadata = {version="*", python="<3.8"}

[tool.poetry.dev-dependencies]
pytest = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
